Changes to incorporate TFT Simulator from TFTMuZeroAgent

+++ core/utils.py
@@ -310,7 +310,7 @@ def prepare_observation_lst(observation_lst):
     batch, stack num, width, height, channel
     """
     # B, S, W, H, C
-    observation_lst = np.array(observation_lst, dtype=np.uint8)
+    observation_lst = np.array(observation_lst)
     observation_lst = np.moveaxis(observation_lst, -1, 2)
 
     shape = observation_lst.shape
